+++ lib/src/batch_instance.dart
@@ -1,0 +1,26 @@
+// Copyright (c) 2022, Kato Shinya. All rights reserved.
+// Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+// Project imports:
+import 'package:batch/src/batch_status.dart';
+
+class BatchInstance {
+  /// The internal constructor.
+  BatchInstance._internal();
+
+  /// Returns the singleton instance of [BatchInstance].
+  static BatchInstance get instance => _singletonInstance;
+
+  /// The singleton instance of this [BatchInstance].
+  static final _singletonInstance = BatchInstance._internal();
+
+  /// The batch status
+  BatchStatus _status = BatchStatus.pending;
+
+  /// Updates batch status to [status].
+  void updateStatus(final BatchStatus status) => _status = status;
+
+  /// Returns true if this batch application is shutting down, otherwise false.
+  bool get isShuttingDown => _status == BatchStatus.shuttingDown;
+}