+++ lib/src/batch_status.dart
@@ -1,0 +1,17 @@
+// Copyright (c) 2022, Kato Shinya. All rights reserved.
+// Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+enum BatchStatus {
+  /// Pending
+  pending,
+
+  /// Running
+  running,
+
+  /// Shutting down
+  shuttingDown,
+
+  /// Shut down
+  shutDown,
+}